requirements.txt
google.generativeai
streamlit
pytube
openai-whisper
ffmpeg-python
semchunk